feat(EMS-2367): renamed function to map
