+++ src/ui/server/middleware/insurance/map-total-contract-value-over-threshold/index.ts
@@ -2,14 +2,15 @@ import { Application } from '../../../../types';
 import { TOTAL_CONTRACT_VALUE } from '../../../constants';
 
 /**
- * totalContractValueOverThreshold
+ * mapTotalContractValueOverThreshold
+ * maps totalContractValue
  * checks if eligibility totalContractValue is above or below threshold
  * if below, then sets application totalContractValueOverThreshold to false
  * if above, then sets application totalContractValueOverThreshold to true
  * @param {Application} application
  * @returns {Application} modifiedApplication
  */
-const totalContractValueOverThreshold = (application: Application) => {
+const mapTotalContractValueOverThreshold = (application: Application) => {
   const modifiedApplication = application;
 
   if (application?.eligibility?.totalContractValue?.value === TOTAL_CONTRACT_VALUE.MORE_THAN_250K.VALUE) {
@@ -21,4 +22,4 @@ const totalContractValueOverThreshold = (application: Application) => {
   return modifiedApplication;
 };
 
-export default totalContractValueOverThreshold;
+export default mapTotalContractValueOverThreshold;
